requirements.txt
django
djangorestframework
markdown
django-filter
drf-yasg
django-environ
psycopg2-binary
pypdf
regex
gunicorn==21.2.0
pdfplumber
python-docx
django-cors-headers
djangorestframework-simplejwt
nltk
pandas
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
sentence-transformers
pinecone-client
